requirements.txt
flask
whisper
python-docx
ffmpeg-python
torch
